svg/geometry/svg-image-with-referred-svg-image-no-intrinsic-size.html

<!DOCTYPE html>
<title><svg:image> 'auto' sizing </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://svgwg.org/svg2-draft/single-page.html#geometry-Sizing"/>
<link rel="help" href="https://svgwg.org/svg2-draft/single-page.html#embedded-Placement"/>
<svg height="0">
  <image width="60" height="60" xlink:href="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='100%' height='100%'></svg>" style='height:auto;'/>
  <image width="60" height="60" xlink:href="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='100%' height='100%'></svg>" style='width:auto;'/>
</svg>
<script>
var expectedBoxes = [
  { dimensions: [60, 150], description: "100%x100% referred svg image, attributes width='60' height='60' and CSS 'height:auto'"},
  { dimensions: [300, 60], description: "100%x100% referred svg image, attributes width='60' height='60' and CSS 'width:auto'"}
];

var images = document.getElementsByTagName('image');
for (var i = 0, length = images.length; i < length; ++i) {
     async_test(function(t) {
        var image = images[i];
        image.expectedBox = expectedBoxes[i];
        image.onload = t.step_func_done(function() {
            var rectBBox = image.getBBox();
            assert_equals(rectBBox.width, image.expectedBox.dimensions[0]);
            assert_equals(rectBBox.height, image.expectedBox.dimensions[1]);
        });
     }, document.title + ' with ' + expectedBoxes[i].description);
}
</script>
